Upgraded base

+++ hath-base/src/main/java/hath/base/CacheHandler.java
@@ -23,6 +23,7 @@ along with Hentai@Home.  If not, see <http://www.gnu.org/licenses/>.
 
 package hath.base;
 
+import java.lang.Thread;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -42,16 +43,14 @@ public class CacheHandler {
 	private int cacheCount = 0, lruClearPointer = 0, lruSkipCheckCycle = 0, pruneAggression = 1;
 	private long cacheSize = 0;
 	private boolean cacheLoaded = false;
-	private Settings settings;
 
-	public CacheHandler(HentaiAtHomeClient client, Settings settings) throws java.io.IOException {
+	public CacheHandler(HentaiAtHomeClient client) throws java.io.IOException {
 		this.client = client;
-		this.settings = settings;
 
-		cachedir = settings.getCacheDir();
+		cachedir = Settings.getCacheDir();
 
 		// delete orphans from the temp dir
-		for (File tmpfile : settings.getTempDir().listFiles()) {
+		for(File tmpfile : Settings.getTempDir().listFiles()) {
 			if(tmpfile.isFile()) {
 				// some silly people might set the data and/or log dir to the same as the temp dir
 				if(!tmpfile.getName().startsWith("log_") && !tmpfile.getName().startsWith("pcache_") && !tmpfile.getName().equals("client_login")) {
@@ -66,7 +65,7 @@ public class CacheHandler {
 
 		boolean fastStartup = false;
 
-		if (!settings.isRescanCache()) {
+		if(!Settings.isRescanCache()) {
 			Out.info("CacheHandler: Attempting to load persistent cache data...");
 
 			if(loadPersistentData()) {
@@ -97,9 +96,9 @@ public class CacheHandler {
 			cacheSize = 0;
 
 			// this is a map with the static ranges in the cache as key and the oldest lastModified file timestamp for every range as value. this is used to find old files to delete if the cache fills up.
-			staticRangeOldest = new Hashtable<String, Long>((int) (settings.getStaticRangeCount() * 1.5));
+			staticRangeOldest = new Hashtable<String,Long>((int) (Settings.getStaticRangeCount() * 1.5));
 
-			if (!settings.isUseLessMemory()) {
+			if(!Settings.isUseLessMemory()) {
 				lruCacheTable = new short[MEMORY_TABLE_ELEMENTS];
 			}
 
@@ -113,20 +112,16 @@ public class CacheHandler {
 		if(!recheckFreeDiskSpace()) {
 			// note: if the client ends up being starved on disk space with static ranges assigned, it will cause a major loss of trust.
 			client.setFastShutdown();
-			client.dieWithError(
-					"The storage device does not have enough space available to hold the given cache size.\nFree up space for H@H, or reduce the cache size from the H@H settings page:\nhttps://e-hentai.org/hentaiathome.php?cid="
-							+ settings.getClientID());
+			client.dieWithError("The storage device does not have enough space available to hold the given cache size.\nFree up space for H@H, or reduce the cache size from the H@H settings page:\nhttps://e-hentai.org/hentaiathome.php?cid=" + Settings.getClientID());
 		}
 
-		if (cacheCount < 1 && settings.getStaticRangeCount() > 20) {
+		if(cacheCount < 1 && Settings.getStaticRangeCount() > 20) {
 			// note: if the client is started with an empty cache and many static ranges assigned, it will cause a major loss of trust.
 			client.setFastShutdown();
-			client.dieWithError(
-					"This client has static ranges assigned to it, but the cache is empty. Check file permissions and file system integrity.\nIf the cache has been deleted or is otherwise lost, you have to manually reset your static ranges from the H@H settings page.\nhttps://e-hentai.org/hentaiathome.php?cid="
-							+ settings.getClientID());
+			client.dieWithError("This client has static ranges assigned to it, but the cache is empty. Check file permissions and file system integrity.\nIf the cache has been deleted or is otherwise lost, you have to manually reset your static ranges from the H@H settings page.\nhttps://e-hentai.org/hentaiathome.php?cid=" + Settings.getClientID());
 		}
 
-		long cacheLimit = settings.getDiskLimitBytes();
+		long cacheLimit = Settings.getDiskLimitBytes();
 
 		if(cacheSize > cacheLimit) {
 			Out.info("CacheHandler: We are over the cache limit, pruning until the limit is met");
@@ -149,15 +144,15 @@ public class CacheHandler {
 	}
 
 	private File getPersistentLRUFile() {
-		return new File(settings.getDataDir(), "pcache_lru");
+		return new File(Settings.getDataDir(), "pcache_lru");
 	}
 
 	private File getPersistentInfoFile() {
-		return new File(settings.getDataDir(), "pcache_info");
+		return new File(Settings.getDataDir(), "pcache_info");
 	}
 
 	private File getPersistentAgesFile() {
-		return new File(settings.getDataDir(), "pcache_ages");
+		return new File(Settings.getDataDir(), "pcache_ages");
 	}
 
 	private boolean loadPersistentData() {
@@ -214,7 +209,7 @@ public class CacheHandler {
 				staticRangeOldest = (Hashtable<String,Long>) readCacheObject(getPersistentAgesFile(), agesHash);
 				Out.info("CacheHandler: Loaded static range ages");
 
-				if (!settings.isUseLessMemory()) {
+				if(!Settings.isUseLessMemory()) {
 					lruCacheTable = (short[]) readCacheObject(getPersistentLRUFile(), lruHash);
 					Out.info("CacheHandler: Loaded LRU cache");
 				}
@@ -300,14 +295,15 @@ public class CacheHandler {
 		Out.info("CacheHandler: Cache cleanup pass..");
 
 		File[] l1dirs = Tools.listSortedFiles(cachedir);
-		int checkedCounter = 0, checkedCounterPct = 0;
+		
+		if(l1dirs == null) {
+			client.dieWithError("CacheHandler: Unable to access " + cachedir + "; check permissions and I/O errors.");
+		}
 
 		// this sanity check can be tightened up when 1.2.6 is EOL and everyone have upgraded to the two-level cache tree
-		// if(l1dirs.length > settings.getStaticRangeCount()) {
-		if (l1dirs.length > 5 && settings.getStaticRangeCount() == 0) {
-			Out.warning("WARNING: There are " + l1dirs.length
-					+ " directories in the cache directory, but the server has only assigned us "
-					+ settings.getStaticRangeCount() + " static ranges.");
+		//if(l1dirs.length > Settings.getStaticRangeCount()) {
+		if(l1dirs.length > 5 && Settings.getStaticRangeCount() == 0) {
+			Out.warning("WARNING: There are " + l1dirs.length + " directories in the cache directory, but the server has only assigned us " + Settings.getStaticRangeCount() + " static ranges.");
 			Out.warning("If this is NOT expected, please close H@H with Ctrl+C or Program -> Shutdown H@H before this timeout expires.");
 			Out.warning("Waiting 30 seconds before proceeding with cache cleanup...");
 
@@ -321,18 +317,25 @@ public class CacheHandler {
 			return;
 		}
 
+		int checkedCounter = 0, checkedCounterPct = 0;
+
 		for(File l1dir : l1dirs) {
 			// time to take out the trash
 			System.gc();
-
+			
 			if(!l1dir.isDirectory()) {
 				l1dir.delete();
 				continue;
 			}
 
 			File[] l2dirs = Tools.listSortedFiles(l1dir);
+			
+			if(l2dirs == null) {
+				Out.warning("CacheHandler: Unable to access " + l1dir + "; check permissions and I/O errors.");
+				continue;
+			}
 
-			if(l2dirs == null || l2dirs.length == 0) {
+			if(l2dirs.length == 0) {
 				l1dir.delete();
 				continue;
 			}
@@ -349,7 +352,7 @@ public class CacheHandler {
 					Out.debug("CacheHandler: The file " + l2dir + " was not recognized.");
 					l2dir.delete();
 				}
-				else if (!settings.isStaticRange(hvFile.getFileid())) {
+				else if( !Settings.isStaticRange(hvFile.getFileid()) ) {
 					Out.debug("CacheHandler: The file " + l2dir + " was not in an active static range.");
 					l2dir.delete();
 				}
@@ -373,8 +376,6 @@ public class CacheHandler {
 	}
 
 	private void startupInitCache() {
-		long recentlyAccessedCutoff = System.currentTimeMillis() - 604800000;
-
 		// update actions:
 		// staticRangeOldest	- add oldest modified timestamp for every static range
 		// addFileToActiveCache	- increments cacheCount and cacheSize
@@ -384,7 +385,7 @@ public class CacheHandler {
 		FileValidator validator = null;
 		int printFreq;
 
-		if (settings.isVerifyCache()) {
+		if(Settings.isVerifyCache()) {
 			Out.info("CacheHandler: Loading cache with full file verification. Depending on the size of your cache, this can take a long time.");
 			validator = new FileValidator();
 			printFreq = 1000;
@@ -394,6 +395,7 @@ public class CacheHandler {
 			printFreq = 10000;
 		}
 		
+		long recentlyAccessedCutoff = System.currentTimeMillis() - 604800000;
 		int foundStaticRanges = 0;
 
 		// cache register pass
@@ -401,8 +403,16 @@ public class CacheHandler {
 			if(!l1dir.isDirectory()) {
 				continue;
 			}
+			
+			File[] l2dirs = Tools.listSortedFiles(l1dir);
 
-			for(File l2dir : Tools.listSortedFiles(l1dir)) {
+			if(l2dirs == null) {
+				// we already warned about level 1 issues in startupCacheCleanup
+				//Out.warning("CacheHandler: Unable to access " + l1dir + "; check permissions and I/O errors.");
+				continue;
+			}
+
+			for(File l2dir : l2dirs) {
 				// the garbage, it must be collected
 				System.gc();
 
@@ -411,6 +421,11 @@ public class CacheHandler {
 				}
 
 				File[] files = Tools.listSortedFiles(l2dir);
+				
+				if(files == null) {
+					Out.warning("CacheHandler: Unable to access " + l2dir + "; check permissions and I/O errors.");
+					continue;
+				}
 
 				if(files.length == 0) {
 					l2dir.delete();
@@ -430,7 +445,7 @@ public class CacheHandler {
 						Out.debug("CacheHandler: The file " + cfile + " was corrupt.");
 						cfile.delete();
 					}
-					else if (!settings.isStaticRange(hvFile.getFileid())) {
+					else if( !Settings.isStaticRange(hvFile.getFileid()) ) {
 						Out.debug("CacheHandler: The file " + cfile + " was not in an active static range.");
 						cfile.delete();
 					}
@@ -474,7 +489,7 @@ public class CacheHandler {
 		}
 
 		long wantFree = 104857600;
-		long cacheLimit = settings.getDiskLimitBytes();
+		long cacheLimit = Settings.getDiskLimitBytes();
 		long bytesToFree = 0;
 
 		if(cacheSize > cacheLimit) {
@@ -486,7 +501,7 @@ public class CacheHandler {
 
 		Out.debug("CacheHandler: Checked cache space (cacheSize=" + cacheSize + ", cacheLimit=" + cacheLimit + ", cacheFree=" + (cacheLimit - cacheSize) + ")");
 
-		if (bytesToFree > 0 && cacheCount > 0 && settings.getStaticRangeCount() > 0) {
+		if(bytesToFree > 0 && cacheCount > 0 && Settings.getStaticRangeCount() > 0) {
 			String pruneStaticRange = null;
 			long nowtime = System.currentTimeMillis();
 			long oldestRangeAge = nowtime;
@@ -576,14 +591,14 @@ public class CacheHandler {
 		// realistically, this is almost certainly unnecessary since the 1.3.2 pruner was added, but it doesn't hurt to have it just in case
 		pruneAggression = bytesToFree > 10485760 ? (int) (bytesToFree / 10485760) : 1;
 
-		if (settings.isSkipFreeSpaceCheck()) {
+		if(Settings.isSkipFreeSpaceCheck()) {
 			Out.debug("CacheHandler: Disk free space check is disabled.");
 			return true;
 		}
 		else {
 			long diskFreeSpace = cachedir.getFreeSpace();
 
-			if (diskFreeSpace < Math.max(settings.getDiskMinRemainingBytes(), wantFree)) {
+			if(diskFreeSpace < Math.max(Settings.getDiskMinRemainingBytes(), wantFree)) {
 				Out.warning("CacheHandler: Did not meet space constraints: Disk free space limit reached (" + diskFreeSpace + " bytes free on device)");
 				return false;
 			}
